work around BCM2837 GPIO bugs
The GPIO block has some quirks, probably due to one of two
possibilities:

1) clock synchronizer boundary crossing issues. The size of the
GPIO glitch is about 8ns, which indicates that the GPIO block might
be running around 125MHz. This is slower than the memory bus and
much slower than the CPU, so writes/reads have to be packaged and
cross a sampling boundary. The set/clear paradigm is not atomic
and when a set/clear splits a clock boundary it could result
in glitches. Furthermore, the GPIO block behaves a bit like there
is a pipeline on the read path, the first read seems to be garbage
or stale, so a dummy read is necessary.

2) cache issues. The above behavior could also be explained
possibly by cache coherency problems.

Either way, the resolution is to insert a dummy read to the
GPIO level state before doing either a write or read to the
block; this seems to force the block into a synchronized state,
but it does stall execution for several instruction cycles.